SA 1 - Main screen
- added bottom navigation bar to main screen
